# Work around potential race in PipeWriter