@halter73 2019-05-10
Nothing changed yet. I'm still going to benchmark/profile this.