@halter73 2019-05-11
Our TE benchmark servers were unavailable this afternoon, but I did get some microbenchmark results from Http1WritingBenchmark. I made a new Http1*Large*WritingBenchmark for larger unawaited writes as well. I fixed the old benchmark (which was hanging previously) and pushed the new new benchmark as part of this PR.

I'm not sure it's worth keeping the new benchmark. I can remove it before merging the PR.

I think the big takeaway is the perf impact is negligible and the allocations are unchanged.

## master

### Http1WritingBenchmark

     Method | WithHeaders | Chunked | OnStarting |     Mean |    Error |   StdDev |        Op/s | Allocated |
----------- |------------ |-------- |----------- |---------:|---------:|---------:|------------:|----------:|
 WriteAsync |       False |   False |      Async | 548.2 ns | 10.96 ns | 21.64 ns | 1,824,087.4 |       0 B |
 WriteAsync |       False |   False |       None | 554.4 ns | 10.49 ns | 18.65 ns | 1,803,866.9 |       0 B |
 WriteAsync |       False |   False |       Sync | 555.6 ns | 11.16 ns | 20.40 ns | 1,799,721.0 |       0 B |
 WriteAsync |       False |    True |      Async | 561.4 ns | 11.16 ns | 23.55 ns | 1,781,149.0 |       0 B |
 WriteAsync |       False |    True |       None | 554.6 ns | 11.07 ns | 22.11 ns | 1,803,227.6 |       0 B |
 WriteAsync |       False |    True |       Sync | 564.4 ns | 11.30 ns | 21.50 ns | 1,771,860.8 |       0 B |
 WriteAsync |        True |   False |      Async | 792.8 ns | 15.70 ns | 27.91 ns | 1,261,393.3 |       0 B |
 WriteAsync |        True |   False |       None | 683.5 ns | 13.27 ns | 19.45 ns | 1,463,063.0 |       0 B |
 WriteAsync |        True |   False |       Sync | 706.8 ns | 13.98 ns | 24.84 ns | 1,414,790.9 |       0 B |
 WriteAsync |        True |    True |      Async | 811.5 ns | 14.10 ns | 18.82 ns | 1,232,335.7 |       0 B |
 WriteAsync |        True |    True |       None | 684.4 ns | 13.65 ns | 20.84 ns | 1,461,183.1 |       0 B |
 WriteAsync |        True |    True |       Sync | 729.7 ns | 14.40 ns | 21.55 ns | 1,370,412.3 |       0 B |


### Http1LargeWritingBenchmark

                      Method |     Mean |     Error |    StdDev |  Op/s |  Gen 0 | Allocated |
---------------------------- |---------:|----------:|----------:|------:|-------:|----------:|
                  WriteAsync | 2.360 ms | 0.0471 ms | 0.0812 ms | 423.8 | 3.9063 |  198.3 KB |
 WriteSegmentsUnawaitedAsync | 2.109 ms | 0.0420 ms | 0.1005 ms | 474.2 |      - |   5.48 KB |

## This PR


### Http1WritingBenchmark

     Method | WithHeaders | Chunked | OnStarting |     Mean |    Error |   StdDev |        Op/s | Allocated |
----------- |------------ |-------- |----------- |---------:|---------:|---------:|------------:|----------:|
 WriteAsync |       False |   False |      Async | 584.6 ns | 11.63 ns | 29.38 ns | 1,710,563.9 |       0 B |
 WriteAsync |       False |   False |       None | 573.2 ns | 11.47 ns | 20.08 ns | 1,744,692.8 |       0 B |
 WriteAsync |       False |   False |       Sync | 562.2 ns | 11.29 ns | 22.01 ns | 1,778,846.3 |       0 B |
 WriteAsync |       False |    True |      Async | 569.3 ns | 10.78 ns | 12.41 ns | 1,756,619.1 |       0 B |
 WriteAsync |       False |    True |       None | 561.2 ns | 11.16 ns | 22.80 ns | 1,782,038.0 |       0 B |
 WriteAsync |       False |    True |       Sync | 565.1 ns | 11.52 ns | 23.28 ns | 1,769,587.7 |       0 B |
 WriteAsync |        True |   False |      Async | 809.5 ns | 15.01 ns | 14.74 ns | 1,235,369.9 |       0 B |
 WriteAsync |        True |   False |       None | 701.1 ns | 13.79 ns | 23.41 ns | 1,426,337.7 |       0 B |
 WriteAsync |        True |   False |       Sync | 735.4 ns | 14.55 ns | 20.87 ns | 1,359,716.2 |       0 B |
 WriteAsync |        True |    True |      Async | 809.5 ns | 17.90 ns | 15.86 ns | 1,235,302.2 |       0 B |
 WriteAsync |        True |    True |       None | 695.8 ns | 13.73 ns | 22.56 ns | 1,437,196.3 |       0 B |
 WriteAsync |        True |    True |       Sync | 719.4 ns | 14.38 ns | 24.02 ns | 1,390,017.6 |       0 B |

### Http1LargeWritingBenchmark

                      Method |     Mean |     Error |    StdDev |  Op/s | Allocated |
---------------------------- |---------:|----------:|----------:|------:|----------:|
                  WriteAsync | 2.340 ms | 0.0463 ms | 0.0823 ms | 427.3 | 198.24 KB |
 WriteSegmentsUnawaitedAsync | 2.190 ms | 0.0436 ms | 0.0691 ms | 456.6 |      5 KB |

## Summary for all runs
```
BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17763                                                                                                
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical cores and 6 physical cores                                                                    
.NET Core SDK=3.0.100-preview5-011568                                                                                                          
  [Host]     : .NET Core 3.0.0-preview6-27708-04 (CoreCLR 3.0.19.25672, CoreFX 4.700.19.25613), 64bit RyuJIT                                   
  Job-NXGMUT : .NET Core 3.0.0-preview6-27708-04 (CoreCLR 3.0.19.25672, CoreFX 4.700.19.25613), 64bit RyuJIT                                   
                                                                                                                 
```